+++ app/_main/_main-old.py
@@ -20,7 +20,7 @@ import timeit
 import logging
 from modules.with_classes import SeatingArrangement
 from modules.with_heapq import hq_seat_arrange
-from modules.with_standalone import sort_seats, arrange_seats
+from with_sort import sort_seats, arrange_seats
 from modules.with_zip_sorted import zip_seat_arrange
 
 # Constants
