+++ app/logs/previous-runs/class.log
@@ -1,0 +1,94 @@
+<!--main.py-->
+Test run 1:
+[[1, True, True], [2, True, True]]
+Execution time: 8.639999941806309e-05
+
+Test run 2:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00013949999993201345
+
+Test run 3:
+[[1, True, True], [2, True, True]]
+Execution time: 0.0002082999999402091
+
+Test run 4:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00012730000162264332
+
+Test run 5:
+[[1, True, True], [2, True, True]]
+Execution time: 0.0002863000008801464
+
+<!--_functions_main.py-->
+[[1, True, True], [2, True, True]]
+Execution time: 1.1599993740674108e-05
+[[1, True, True], [2, True, True]]
+Execution time: 4.599998646881431e-06
+[[1, True, True], [2, True, True]]
+Execution time: 3.9000005926936865e-06
+[[1, True, True], [2, True, True]]
+Execution time: 1.3500000932253897e-05
+[[1, True, True], [2, True, True]]
+Execution time: 1.0399999155197293e-05
+[[1, True, True], [2, True, True]]
+Execution time: 5.099995178170502e-06
+[[1, True, True], [2, True, True]]
+Execution time: 5.199995939619839e-06
+[[1, True, True], [2, True, True]]
+Execution time: 1.5300000086426735e-05
+[[1, True, True], [2, True, True]]
+Execution time: 1.3600001693703234e-05
+[[1, True, True], [2, True, True]]
+
+<!--_decorators_main.py-->
+Execution time: 7.099995855242014e-06
+Execution time: 0.0007196000005933456
+Execution time: 1.2399999832268804e-05
+Execution time: 9.000003046821803e-06
+Execution time: 1.0399999155197293e-05
+Execution time: 3.0499999411404133e-05
+Execution time: 1.5900004655122757e-05
+Execution time: 9.499999578110874e-06
+Execution time: 1.2600001355167478e-05
+Execution time: 3.1400006264448166e-05
+Execution time: 9.300005331169814e-06
+
+Test run 1:
+[[1, True, True], [2, True, True]]
+Execution time: 0.0001486000037402846
+
+Test run 2:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00011480000102892518
+
+Test run 3:
+[[1, True, True], [2, True, True]]
+Execution time: 5.099999543745071e-05
+
+Test run 4:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00022729999909643084
+
+Test run 5:
+[[1, True, True], [2, True, True]]
+Execution time: 7.870000263210386e-05
+
+Test run 1:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00015330000314861536
+
+Test run 2:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00022190000163391232
+
+Test run 3:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00021470000501722097
+
+Test run 4:
+[[1, True, True], [2, True, True]]
+Execution time: 5.1499984692782164e-05
+
+Test run 5:
+[[1, True, True], [2, True, True]]
+Execution time: 6.329998723231256e-05
