+++ app/logs/previous-runs/hq.log
@@ -1,0 +1,90 @@
+
+Test run 1:
+['BG', 'BG']
+Execution time: 0.00014870000086375512
+
+Test run 2:
+['BG', 'BG']
+Execution time: 0.00010009999823523685
+
+Test run 3:
+['BG', 'BG']
+Execution time: 8.849999721860513e-05
+
+Test run 4:
+['BG', 'BG']
+Execution time: 9.4600000011269e-05
+
+Test run 5:
+['BG', 'BG']
+Execution time: 0.00015630000052624382
+['BG', 'BG']
+Execution time: 4.6799999836366624e-05
+['BG', 'BG']
+Execution time: 1.1500000255182385e-05
+['BG', 'BG']
+Execution time: 1.3000004400964826e-05
+['BG', 'BG']
+Execution time: 1.1299998732283711e-05
+['BG', 'BG']
+Execution time: 8.600000001024455e-06
+['BG', 'BG']
+Execution time: 9.800001862458885e-06
+['BG', 'BG']
+Execution time: 4.350000381236896e-05
+['BG', 'BG']
+Execution time: 1.049999991664663e-05
+['BG', 'BG']
+Execution time: 2.8700000257231295e-05
+['BG', 'BG']
+Execution time: 1.1099997209385037e-05
+Execution time: 1.5600002370774746e-05
+Execution time: 7.88000033935532e-05
+Execution time: 1.5600002370774746e-05
+Execution time: 1.0900002962443978e-05
+Execution time: 1.5600002370774746e-05
+Execution time: 1.3000004400964826e-05
+Execution time: 1.85999961104244e-05
+Execution time: 4.119999357499182e-05
+Execution time: 1.0700001439545304e-05
+Execution time: 9.200004569720477e-06
+
+Test run 1:
+['BG', 'BG']
+Execution time: 0.0001416000013705343
+
+Test run 2:
+['BG', 'BG']
+Execution time: 0.00029379999614320695
+
+Test run 3:
+['BG', 'BG']
+Execution time: 7.750000077066943e-05
+
+Test run 4:
+['BG', 'BG']
+Execution time: 0.00011969999468419701
+
+Test run 5:
+['BG', 'BG']
+Execution time: 9.790000331122428e-05
+
+Test run 1:
+['BG', 'BG']
+Execution time: 6.580000626854599e-05
+
+Test run 2:
+['BG', 'BG']
+Execution time: 0.00024029999622143805
+
+Test run 3:
+['BG', 'BG']
+Execution time: 0.00025209999876096845
+
+Test run 4:
+['BG', 'BG']
+Execution time: 0.00021739999647252262
+
+Test run 5:
+['BG', 'BG']
+Execution time: 0.00022220000391826034
