+++ app/logs/previous-runs/standalone.log
@@ -1,0 +1,90 @@
+
+Test run 1:
+[[1, True, True], [2, True, True]]
+Execution time: 0.0003431999975873623
+
+Test run 2:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00026390000130049884
+
+Test run 3:
+[[1, True, True], [2, True, True]]
+Execution time: 7.959999857121147e-05
+
+Test run 4:
+[[1, True, True], [2, True, True]]
+Execution time: 0.0001019000010273885
+
+Test run 5:
+[[1, True, True], [2, True, True]]
+Execution time: 5.270000110613182e-05
+[[1, True, True], [2, True, True]]
+Execution time: 5.299996701069176e-06
+[[1, True, True], [2, True, True]]
+Execution time: 3.7999998312443495e-06
+[[1, True, True], [2, True, True]]
+Execution time: 1.200000406242907e-05
+[[1, True, True], [2, True, True]]
+Execution time: 4.0000013541430235e-06
+[[1, True, True], [2, True, True]]
+Execution time: 3.7999998312443495e-06
+[[1, True, True], [2, True, True]]
+Execution time: 5.30000397702679e-06
+[[1, True, True], [2, True, True]]
+Execution time: 3.6999990697950125e-06
+[[1, True, True], [2, True, True]]
+Execution time: 2.7999994927085936e-06
+[[1, True, True], [2, True, True]]
+Execution time: 5.200003215577453e-06
+[[1, True, True], [2, True, True]]
+Execution time: 4.800000169780105e-06
+Execution time: 9.499999578110874e-06
+Execution time: 2.710000262595713e-05
+Execution time: 1.0900002962443978e-05
+Execution time: 7.599999662488699e-06
+Execution time: 9.60000033956021e-06
+Execution time: 1.0199997632298619e-05
+Execution time: 8.100003469735384e-06
+Execution time: 6.599999323952943e-06
+Execution time: 7.199996616691351e-06
+Execution time: 7.399998139590025e-06
+
+Test run 1:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00017870000010589138
+
+Test run 2:
+[[1, True, True], [2, True, True]]
+Execution time: 0.0002367999986745417
+
+Test run 3:
+[[1, True, True], [2, True, True]]
+Execution time: 7.069999992381781e-05
+
+Test run 4:
+[[1, True, True], [2, True, True]]
+Execution time: 7.729999924777076e-05
+
+Test run 5:
+[[1, True, True], [2, True, True]]
+Execution time: 8.420000085607171e-05
+
+Test run 1:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00018030000501312315
+
+Test run 2:
+[[1, True, True], [2, True, True]]
+Execution time: 0.0001066999975591898
+
+Test run 3:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00011190000805072486
+
+Test run 4:
+[[1, True, True], [2, True, True]]
+Execution time: 0.00011530000483617187
+
+Test run 5:
+[[1, True, True], [2, True, True]]
+Execution time: 5.639999289996922e-05
