+++ app/logs/previous-runs/zip.log
@@ -1,0 +1,90 @@
+
+Test run 1:
+['B', 'B']
+Execution time: 8.269999671028927e-05
+
+Test run 2:
+['B', 'B']
+Execution time: 0.00017519999892101623
+
+Test run 3:
+['B', 'B']
+Execution time: 0.0001650000012887176
+
+Test run 4:
+['B', 'B']
+Execution time: 0.00013380000018514693
+
+Test run 5:
+['B', 'B']
+Execution time: 9.069999941857532e-05
+['B', 'B']
+Execution time: 9.10000380827114e-06
+['B', 'B']
+Execution time: 1.2999997125007212e-05
+['B', 'B']
+Execution time: 1.0700001439545304e-05
+['B', 'B']
+Execution time: 8.999995770864189e-06
+['B', 'B']
+Execution time: 1.1300006008241326e-05
+['B', 'B']
+Execution time: 1.0000003385357559e-05
+['B', 'B']
+Execution time: 9.199997293762863e-06
+['B', 'B']
+Execution time: 2.0200001017656177e-05
+['B', 'B']
+Execution time: 1.3500000932253897e-05
+['B', 'B']
+Execution time: 5.390000296756625e-05
+Execution time: 2.0699997548945248e-05
+Execution time: 1.5199999324977398e-05
+Execution time: 2.42000023717992e-05
+Execution time: 1.1800002539530396e-05
+Execution time: 1.0399999155197293e-05
+Execution time: 1.7700003809295595e-05
+Execution time: 1.6099998902063817e-05
+Execution time: 1.309999788645655e-05
+Execution time: 3.699999797390774e-05
+Execution time: 1.0899995686486363e-05
+
+Test run 1:
+['B', 'B']
+Execution time: 7.819999882485718e-05
+
+Test run 2:
+['B', 'B']
+Execution time: 0.00020930000027874485
+
+Test run 3:
+['B', 'B']
+Execution time: 7.750000077066943e-05
+
+Test run 4:
+['B', 'B']
+Execution time: 0.0002395999981672503
+
+Test run 5:
+['B', 'B']
+Execution time: 9.819999831961468e-05
+
+Test run 1:
+['B', 'B']
+Execution time: 6.160000339150429e-05
+
+Test run 2:
+['B', 'B']
+Execution time: 0.00024349999148398638
+
+Test run 3:
+['B', 'B']
+Execution time: 8.05000017862767e-05
+
+Test run 4:
+['B', 'B']
+Execution time: 8.759999764151871e-05
+
+Test run 5:
+['B', 'B']
+Execution time: 5.370000144466758e-05
