+++ tests/test_main.py
@@ -50,23 +50,28 @@ _summary_
       something to consider for more complex applications.
 """
 
+from typing import Any
 import pytest
-from modules.with_classes import SeatingArrangement
-from modules.with_heapq import hq_seat_arrange
-from modules.with_standalone import sort_seats, arrange_seats
-from modules.with_zip_sorted import zip_seat_arrange
+from app.modules.with_classes import SeatingArrangement
+from app.modules.with_heapq import hq_seat_arrange
+from app.modules.with_sort import sort_seats, arng_seats
+from app.modules.with_zip_sorted import zip_seat_arrange
 
 # List of all the methods for easy testing
 methods = [
     lambda N, S, W: SeatingArrangement(N, S, W).arrange_seats(),
     hq_seat_arrange,
-    lambda N, S, W: arrange_seats(N, S, sort_seats(W)),
+    lambda N, S, W: arng_seats(N, S, sort_seats(W)),
     zip_seat_arrange
 ]
 
 #------------------- 1. Basic Functionality Tests ------------------------------
 @pytest.mark.parametrize("method", methods)
-def test_basic_functionality(method):
+def test_basic_functionality(method: Any):
+    """
+    Test each method with a variety of inputs to ensure they produce the
+    expected outputs.
+    """
     N = 2
     S = '0011'
     W = [2, 1]
@@ -76,14 +81,16 @@ def test_basic_functionality(method):
 
 #------------------------- 2. Edge Cases ----------------------------
 @pytest.mark.parametrize("method", methods)
-def test_empty_hall(method):
+def test_empty_hall(method: Any):
+    """Test with an empty hall (no rows)."""
     N = 0
     S = ''
     W = []
     assert method(N, S, W) == []
 
 @pytest.mark.parametrize("method", methods)
-def test_only_boys(method):
+def test_only_boys(method: Any):
+    """Test with a sequence containing only boys or only girls."""
     N = 2
     S = '00'
     W = [2, 1]
@@ -92,7 +99,8 @@ def test_only_boys(method):
     assert method(N, S, W) == expected_output
 
 @pytest.mark.parametrize("method", methods)
-def test_only_girls(method):
+def test_only_girls(method: Any):
+    """Test with a sequence containing only boys or only girls."""
     N = 2
     S = '11'
     W = [2, 1]
@@ -101,7 +109,8 @@ def test_only_girls(method):
     assert method(N, S, W) == expected_output
 
 @pytest.mark.parametrize("method", methods)
-def test_same_width(method):
+def test_same_width(method: Any):
+    """Test with all rows having the same width."""
     N = 2
     S = '0011'
     W = [2, 2]
@@ -111,7 +120,8 @@ def test_same_width(method):
 
 # ------------------------ 3. Invalid Inputs ----------------------------------
 @pytest.mark.parametrize("method", methods)
-def test_negative_rows(method):
+def test_negative_rows(method: Any):
+    """Test with negative numbers of rows."""
     N = -2
     S = '0011'
     W = [2, 1]
@@ -119,7 +129,10 @@ def test_negative_rows(method):
         method(N, S, W)
 
 @pytest.mark.parametrize("method", methods)
-def test_non_binary_sequence(method):
+def test_non_binary_sequence(method: Any):
+    """Test with non-binary sequences (i.e., sequences containing characters
+    other than '0' and '1').
+    """
     N = 2
     S = '0021'
     W = [2, 1]
@@ -127,7 +140,8 @@ def test_non_binary_sequence(method):
         method(N, S, W)
 
 @pytest.mark.parametrize("method", methods)
-def test_mismatched_lengths(method):
+def test_mismatched_lengths(method: Any):
+    """Test with mismatched lengths of sequence and width lists."""
     N = 2
     S = '0011'
     W = [2]
@@ -135,7 +149,8 @@ def test_mismatched_lengths(method):
         method(N, S, W)
 
 @pytest.mark.parametrize("method", methods)
-def test_negative_width(method):
+def test_negative_width(method: Any):
+    """Test with negative or zero widths."""
     N = 2
     S = '0011'
     W = [-2, 1]
@@ -146,7 +161,11 @@ def test_negative_width(method):
 # This is more for benchmarking and might not have a clear 'expected' result.
 # It's more about ensuring the methods can handle large inputs without errors.
 @pytest.mark.parametrize("method", methods)
-def test_large_input(method):
+def test_large_input(method: Any):
+    """
+    Test with a large number of rows and a long sequence to measure the
+    performance of each method.
+    """
     N = 1000
     S = '0' * 500 + '1' * 500
     W = [i for i in range(1, 1001)]
@@ -155,6 +174,11 @@ def test_large_input(method):
 
 # ------------------------- 5. Consistency Tests -----------------------------
 def test_consistency():
+    """
+    - Ensure that all methods produce the same output for the same input.
+    - This can be done by comparing the outputs of the different methods for a
+      variety of inputs.
+    """
     N = 2
     S = '0011'
     W = [2, 1]
